requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
pypdf==5.0.0
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
sentence-transformers==3.1.0
faiss-cpu
rank-bm25==0.2.2
groq==0.11.0
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==13.0
pydantic==2.9.0
numpy
tiktoken==0.7.0
python-pptx==1.0.2
